Split up HTTP `DurationChart` into two components
